requirements.txt
opencv-python
pillow
roboflow
ultralytics